requirements.txt
Django==5.2.5
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.9
PyJWT==2.10.1
sqlparse==0.5.3